Staging: line6/midi.c: Fixed call of obsolete function strict_strtoul
The obsolete function strict_strtoul should be replaced by the kstrto*
functions.

In this context kstrtou16 should be use, as midi_mask_receive is only
used as unsigned short. All corresponding datatypes were adapted
accordingly.

+++ drivers/staging/line6/midi.h
@@ -62,7 +62,7 @@ struct snd_line6_midi {
 	/**
 		 Bit mask for input MIDI channels.
 	*/
-	int midi_mask_receive;
+	unsigned short midi_mask_receive;
 
 	/**
 		 Buffer for incoming MIDI stream.
